Merge Product into Table
Since the Product component is very tightly coupled with the Table,
it didn't really make sense to have it as it's own component.

I'd prefer it if the module boundaries were at the level of reusable
solutions. That is, each module provides a complete solution to some
problem, and has only explicit dependencies/requirements.